Rename: auth -> user
이유: 백엔드 API 명세에 로그인 엔드포인트와 유저 관련 엔드포인트가 동일하므로 auth 와 user 를 분리하지 않고 user로 통일
