+++ src/apis/user.api.ts
@@ -1,7 +1,7 @@
 import { httpClient } from "./http.api";
 
 export const createLogin = async (email: string, password: string) => {
-    const response = await httpClient.post("/login", { email, password });
+    const response = await httpClient.post("/user/login", { email, password });
     return response.data;
 };
