+++ src/mock/browser.mock.ts
@@ -1,6 +1,6 @@
 import { HttpHandler } from "msw";
 import { setupWorker } from "msw/browser";
-import { authHandlers } from "./auth.mock";
+import { authHandlers } from "./user.mock";
 
 const handlers: HttpHandler[] = [...authHandlers];
